+++ packages/expo-status-bar/CHANGELOG.md
@@ -12,6 +12,8 @@
 
 ### 💡 Others
 
+- Minimize modules.
+
 ## 1.12.1 — 2024-04-23
 
 _This version does not introduce any user-facing changes._
