+++ packages/expo-status-bar/package.json
@@ -2,7 +2,7 @@
   "name": "expo-status-bar",
   "version": "1.12.1",
   "description": "Provides the same interface as the React Native StatusBar API, but with slightly different defaults to work great in Expo environments.",
-  "main": "src/StatusBar.ts",
+  "main": "src/StatusBar",
   "types": "build/StatusBar.d.ts",
   "scripts": {
     "build": "expo-module build",
