--- packages/expo-status-bar/src/styleToBarStyle.ts
@@ -1,28 +1,0 @@
-import { Appearance, ColorSchemeName } from 'react-native';
-
-import type { StatusBarStyle } from './StatusBar.types';
-
-function getColorScheme() {
-  if (Appearance) {
-    return Appearance.getColorScheme();
-  }
-  return 'light';
-}
-
-export default function styleToBarStyle(
-  style: StatusBarStyle = 'auto',
-  colorScheme: ColorSchemeName = getColorScheme()
-): 'light-content' | 'dark-content' {
-  if (!colorScheme) {
-    colorScheme = 'light';
-  }
-
-  let resolvedStyle = style;
-  if (style === 'auto') {
-    resolvedStyle = colorScheme === 'light' ? 'dark' : 'light';
-  } else if (style === 'inverted') {
-    resolvedStyle = colorScheme === 'light' ? 'light' : 'dark';
-  }
-
-  return resolvedStyle === 'light' ? 'light-content' : 'dark-content';
-}
